+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - New `/lab/project-aidah` page
+- Button "Terapkan" on bottom sheet in "/lab/murajaah-at-taisir" page for better UX
 
 ## [0.17.0] - 2026-02-23